Move service specific models to separate package
