+++ maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/annotation/DefaultHTTPErrorResponse.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.http.MediaType;
-import uk.gov.justice.laa.crime.orchestration.dto.maat.ErrorDTO;
+import uk.gov.justice.laa.crime.orchestration.dto.ErrorDTO;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
