+++ maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/controller/HardshipController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.*;
 import uk.gov.justice.laa.crime.orchestration.annotation.DefaultHTTPErrorResponse;
 import uk.gov.justice.laa.crime.orchestration.dto.maat.ApplicationDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.maat.HardshipReviewDTO;
-import uk.gov.justice.laa.crime.orchestration.dto.maat.WorkflowRequest;
+import uk.gov.justice.laa.crime.orchestration.dto.WorkflowRequest;
 import uk.gov.justice.laa.crime.orchestration.service.HardshipOrchestrationService;
 
 import static uk.gov.justice.laa.crime.commons.common.Constants.LAA_TRANSACTION_ID;
