+++ maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/dto/WorkflowRequest.java
@@ -1,9 +1,11 @@
-package uk.gov.justice.laa.crime.orchestration.dto.maat;
+package uk.gov.justice.laa.crime.orchestration.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import uk.gov.justice.laa.crime.orchestration.dto.maat.ApplicationDTO;
+import uk.gov.justice.laa.crime.orchestration.dto.maat.UserDTO;
 
 @Data
 @Builder
